refactor: exam 페이지 코드리뷰 추가 수정(#52)
- fix : 코드충돌 해결
