+++ src/constants/urls.ts
@@ -1,10 +1,5 @@
 export const ROUTES_PATHS = {
   MAIN: '/',
+  EXAM: '/exams',
   NOT_FOUND: '*',
-  EXAM: '/exams/exams',
-}
-
-export const PAGE_TITLES: Record<string, string> = {
-  [ROUTES_PATHS.MAIN]: '쪽지시험 관리',
-  [ROUTES_PATHS.EXAM]: '쪽지시험 관리',
 }
