--- src/features/exam/index.ts
@@ -1,11 +1,0 @@
-// exam
-export { default as FilterSection } from './FilterSection'
-export { default as EmptyState } from './EmptyState'
-export { default as ExamList } from './ExamList'
-export { default as ExamPage } from '../../pages/ExamManagementPage.tsx'
-export { ExamColumns } from './examConfig.tsx'
-export { mockExamList } from '../../mocks/mockExamList.ts'
-export { default as ExamCreate } from './ExamCreate'
-
-// utils
-export { createExamDropdown } from './utils/createExamDropdown.ts'
