+++ src/features/exams/components/EmptyState.tsx
@@ -8,6 +8,8 @@ type EmptyStateProps = {
  * 쪽지시험 데이터가 없을 때 노출되는 빈 상태 컴포넌트
  * - 시험이 등록되지 않은 경우 사용자에게 안내 메시지를 표시
  * - 시험 생성 버튼을 통해 생성 플로우로 유도
+ * - 버튼 클릭 시 'onButtonClick' 콜백을 실행하여
+ *   상위 컴포넌트에서 정의한 로직 트리거
  */
 export default function EmptyState({ onButtonClick }: EmptyStateProps) {
   return (
