+++ src/features/exams/components/examConfig.tsx
@@ -2,7 +2,7 @@ import type { ColumnDef } from '@tanstack/react-table'
 
 import { Button } from '@components'
 
-import type { Exam } from './types'
+import type { Exam } from '../types'
 
 /**
  * 쪽지시험 목록 테이블 컬럼 설정
