+++ src/features/exams/story/ExamCreate.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react-vite'
 
-import ExamCreate from '../ExamCreate'
+import { ExamCreate } from '@exams'
 
 const meta = {
   title: 'Modals/ExamCreate',
