+++ src/pages/ExamManagementPage.tsx
@@ -1,18 +1,11 @@
+import { EmptyState, ExamList, FilterSection } from '@exams'
 import { COURSE_LIST_DROPDOWN, SUBJECT_LIST_DROPDOWN } from '@mocks'
 import { useState } from 'react'
 
-import type {
-  DropdownConfig,
-  Exam,
-  ExamPageProps,
-} from '../features/exam/types'
-
-import EmptyState from '../features/exam/EmptyState'
-import ExamList from '../features/exam/ExamList'
-import FilterSection from '../features/exam/FilterSection'
+import type { DropdownConfig, Exam } from '../features/exams/types'
 
 // 드롭다운 설정
-const ExAM_DROPDOWNS: DropdownConfig[] = [
+const EXAM_DROPDOWNS: DropdownConfig[] = [
   { key: 'course', items: COURSE_LIST_DROPDOWN, placeholder: '과정' },
   { key: 'subject', items: SUBJECT_LIST_DROPDOWN, placeholder: '과목' },
 ]
@@ -22,16 +15,15 @@ const ExAM_DROPDOWNS: DropdownConfig[] = [
  * - 필터, 검색, 시험 목록/빈 상태를 관리하는 컨테이너 컴포넌트
  * - 시험 데이터 유무에 따라 EmptyState 또는 ExamList 렌더링
  */
-export default function ExamManagementPage({
-  initialData = [],
-}: ExamPageProps) {
+export default function ExamManagementPage() {
+  // TODO: API 연동 시 useQuery 등으로 내부에서 시험 목록 fetch 예정
+  const [data, _setData] = useState<Exam[]>([])
+
   const [filters, setFilters] = useState<Record<string, string>>({
     course: '',
     subject: '',
   })
 
-  // TODO: API 연동 시 서버에서 시험 목록 데이터를 조회하도록 변경 예정
-  const [data, _setData] = useState<Exam[]>(initialData)
   const [search, setSearch] = useState('')
 
   /**
@@ -43,11 +35,19 @@ export default function ExamManagementPage({
     setFilters((prev) => ({ ...prev, [key]: value }))
   }
 
+  /**
+   * 검색 버튼 클릭 핸들러
+   * - 현재 필터와 검색어를 기반으로 시험 목록 조회
+   */
   const handleSearch = () => {
     // eslint-disable-next-line no-console
     console.log('조회', { filters, search })
   }
 
+  /**
+   * 시험 생성 버튼 클릭 핸들러
+   * - 시험 생성 페이지로 이동 또는 모달 오픈
+   */
   const handleCreate = () => {
     // eslint-disable-next-line no-console
     console.log('시험 생성하기')
@@ -60,7 +60,7 @@ export default function ExamManagementPage({
 
         <div className="mb-3">
           <FilterSection
-            dropdowns={ExAM_DROPDOWNS}
+            dropdowns={EXAM_DROPDOWNS}
             selectedValues={filters}
             onChangeFilters={handleChangeFilters}
             search={search}
