+++ tsconfig.app.json
@@ -46,7 +46,9 @@
       "@custom-types": ["src/custom-types"],
       "@custom-types/*": ["src/custom-types/*"],
       "@mocks": ["src/mocks"],
-      "@mocks/*": ["src/mocks/*"]
+      "@mocks/*": ["src/features/exam*"],
+      "@exams": ["src/features/exams"],
+      "@exams/*": ["src/features/exams*"]
     }
   },
   "include": ["src"]
